requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.8
dnspython==2.6.1
filelock==3.15.4
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
pycparser==2.22
pymongo==4.8.0
python-dotenv==1.0.1
virtualenv==20.26.3
Werkzeug==3.0.4
